Add Support single binary profiles when CONFIG_PID_IN_CONTEXTIDR is not set
This patch adds support for single binary profiling without
CONFIG_PID_IN_CONTEXTIDR.

I.e, this commit adds a fallback for SPE-based profiling on kernels that
are compiled without the `CONFIG_PID_IN_CONTEXTIDR` option.

Previously, running `perf` on a single application would
fail in this environment because the PID could not be determined from the
SPE packet. This change ensures the profiler can still operate in this
scenario.